requirements.txt
python-dotenv~=1.0.0
telegram
python-telegram-bot~=20.5
openai~=0.27.7
cryptography~=41.0.1